+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.8.8
 
+* `Versions`:
+    * `AppCompat`: `1.3.1` -> `1.4.0`
+    * Android Compile SDK: `31.0.0` -> `32.0.0`
+
 ## 0.8.7
 
 * `Ktor`: